requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
openai==1.3.7
python-dotenv==1.0.0 